fix padded input

+++ backend/src/pytorch/InterpolateArchs/RIFE/rife421IFNET.py
@@ -81,7 +81,7 @@ class Head(nn.Module):
         x3 = self.cnn3(x)
         if feat:
             return [x0, x1, x2, x3]
-        return x3.float()
+        return x3
 
 
 class ResConv(nn.Module):
